Refactor AuthResponse and JwtPayload structures
Updated the `AuthResponse` interface to replace the `user` property in the `result` object with `id`, `name`, and `userName`. Modified the `JwtPayload` interface to change `sub`, `name`, and `userName` to `UserId`, `Name`, and `UserName`. Adjusted the `AuthService` class to handle the new properties and set `currentUserSubject` to the entire `result` object for both login and registration processes.